+++ mex/common/models/resource.py
@@ -5,7 +5,7 @@ from typing import Annotated, ClassVar, Literal
 from pydantic import AfterValidator, Field, computed_field
 
 from mex.common.models.base.extracted_data import ExtractedData
-from mex.common.models.base.merged_item import MergedItem
+from mex.common.models.base.merged_item import MergedItem, PreviewItem
 from mex.common.models.base.model import BaseModel
 from mex.common.models.base.rules import (
     AdditiveRule,
@@ -249,6 +249,17 @@ class MergedResource(BaseResource, MergedItem):
     entityType: Annotated[
         Literal["MergedResource"], Field(alias="$type", frozen=True)
     ] = "MergedResource"
+    identifier: Annotated[MergedResourceIdentifier, Field(frozen=True)]
+
+
+class PreviewResource(
+    _OptionalLists, _SparseLists, _OptionalValues, _SparseValues, PreviewItem
+):
+    """Preview for merging all extracted data and rules for a resource."""
+
+    entityType: Annotated[
+        Literal["PreviewResource"], Field(alias="$type", frozen=True)
+    ] = "PreviewResource"
     identifier: Annotated[MergedResourceIdentifier, Field(frozen=True)]